refactor: move Text component to separate directory
