--- src/content/Text.js
@@ -1,31 +1,0 @@
-import React, {Fragment, useState} from 'react';
-import GalleryPicture from './icons/GalleryPicture';
-import './media-gallery.css';
-
-export const TextAddButton = (props) => {
-  return (
-    <button className="media-gallery-add-button" {...props}>
-      <div className="media-gallery-add-button__picture">
-        <GalleryPicture />
-      </div>
-      <div className="media-gallery-add-button__label">Photo gallery</div>
-    </button>
-  );
-};
-
-export const TextEdit = ({data, onChange}) => {
-  return (
-    <textarea
-      value={(data && data.text) || ''}
-      onChange={(event) => {
-        onChange({text: event.target.value});
-      }}
-    />
-  );
-};
-
-const Text = ({data}) => {
-  return <p>{data.text}</p>;
-};
-
-export default Text;
